requirements.txt
py-rattler
pyyaml
